refactor: filter features 분리
