+++ src/widgets/post/detail-dialog/index.tsx
@@ -1,18 +1,17 @@
 import { useSelectedPost } from "@/entity/post/model/selectedPost"
 import { useDialogStore } from "@/features/dialog/model/store"
 
-import { useSearchQueryStore } from "@/features/filter/search/model/store"
 import { highlightText } from "@/shared/lib/highlightText"
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/shared/ui"
 
 import { CommentList } from "@/widgets/comment/list"
-
+import { useSearchStore } from "@/features/search/model/useSearchStore"
 export const POST_DETAIL_DIALOG = "POST_DETAIL_DIALOG"
 
 export const PostDetailDialog = () => {
   const { isDialogOpen, setDialogOpen } = useDialogStore()
   const { selectedPost } = useSelectedPost()
-  const { searchQuery } = useSearchQueryStore()
+  const { searchQuery } = useSearchStore()
 
   if (!selectedPost) return null
