+++ src/widgets/post/filter/index.tsx
@@ -1,7 +1,7 @@
 import { TagSelect } from "@/features/tag/select"
 
-import { SortBySelect } from "@/features/filter/sort-by"
-import { SortOrderSelector } from "@/features/filter/sort-order"
+import { SortBySelect } from "@/widgets/filter/sort-by-select"
+import { SortOrderSelector } from "@/widgets/filter/sort-order-select"
 import { SearchInput } from "@/features/search"
 
 export const PostFilter = () => {
